Revert "LPS-28891"
This reverts commit fc24905a9c10594c8e85103256c7456a787c5f49.
